Convert float to double

+++ ArduinoNanoRP2040Connect/polar_plotter/safeStatus.h
@@ -42,9 +42,9 @@ public:
   void init();
 
   void writeStatus(const String &key, const String &value);
-  void setMaxRadius(const float value);
-  void setRadiusStepSize(const float value);
-  void setAzimuthStepSize(const float value);
+  void setMaxRadius(const double value);
+  void setRadiusStepSize(const double value);
+  void setAzimuthStepSize(const double value);
   void setMarbleSizeInRadiusSteps(const int value);
   void setCurrentDrawing(const String &value);
   void setCurrentStep(const String &value);
